reconstruct server
