+++ src/Makefile
@@ -5,12 +5,13 @@ CABINET_CLIENT_NAME=cabinet-cli
 
 #Common Dependency
 CABINET_COMMAND_OBJECT=GetCommand.o SetCommand.o DelCommand.o NoMatchCommand.o CommandKeeper.o 
-CABINET_COMMON_HEADER=Command.h Const.h ValueObj.h CabinetLogo.h
+CABINET_COMMON_HEADER=Command.h Const.h ValueObj.h CabinetLogo.h Cabinet.h
 CABINET_COMMON_OBJECT=Log.o Util.o ProtocolStream.o $(CABINET_COMMAND_OBJECT)
 CABINET_DATABASE_OBJECT=DataBase.o StringObj.o
 
 #Target Dependency
-CABINET_SERVER_OBJECT=cabinet-server.o Server.o Client.o EventPoll.o PersistenceFile.o  $(CABINET_COMMON_OBJECT) $(CABINET_DATABASE_OBJECT)
+CABINET_SERVER_OBJECT=cabinet-server.o Cabinet.o Server.o Client.o EventPoll.o PersistenceFile.o \
+						$(CABINET_COMMON_OBJECT) $(CABINET_DATABASE_OBJECT)
 CABINET_CLIENT_OBJECT=cabinet-cli.o CabinetCli.o $(CABINET_COMMON_OBJECT)
 
 #Compile Arguements
