+++ src/cabinet-server.cpp
@@ -2,10 +2,10 @@
 
 int main () 
 {
-    Server server;
-    server.initConfig();
-    server.init();
-    server.onFire();
+    Cabinet *server = new Server();
+    server->initConfig();
+    server->init();
+    server->onFire();
 
     return 0;
 }
